requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
attrs==25.3.0
behave==1.2.6
brotlicffi==1.1.0.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ecdsa==0.19.1
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
MouseInfo==0.1.3
multidict==6.5.0
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.4
playwright==1.54.0
propcache==0.3.2
psutil==7.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pyee==13.0.0
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyNaCl==1.5.0
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dotenv==1.1.1
python-graphql-client==0.4.3
python-telegram-bot==22.3
APScheduler==3.10.4
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.4
selenium==4.35.0
setuptools==78.1.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
types-six==1.17.0.20250515
typing_extensions==4.14.0
undetected-chromedriver==3.5.5
urllib3==2.5.0
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
win-inet-pton==1.1.0
wsproto==1.2.0
yarl==1.20.1
